refactor: extract common abstractions for each connector (#7528)
Approved-By: waruto210
